Enhance model loading and experiment configuration
- Added `dtype` parameter to `ExperimentConfig` for model weight precision.
- Updated `load_model` function to accept `dtype` for improved memory efficiency.
- Introduced new command-line arguments for `--dtype` and `--rho-ladder` in `run_all.py` to facilitate flexible experiment setups.
- Enhanced argument parsing in `run_experiment.py` to include dtype options, ensuring consistent model configuration across experiments.

+++ src/secondary_experiments/experiment.py
@@ -204,7 +204,7 @@ def run_with_llm(config: ExperimentConfig = DEFAULT_CONFIG) -> list[dict]:
     )
     cache = CacheBaseline(alpha=config.alpha)
 
-    model = load_model(config.model_name, device=config.device)
+    model = load_model(config.model_name, device=config.device, dtype=config.dtype)
     token_map = build_token_map(model)
 
     rows: list[dict] = []
